rl implementation + run3b

+++ unsloth_compiled_cache/Conv2d.py
@@ -62,7 +62,7 @@ import torch
 import torch.nn as nn
 from torch.nn import functional as F
 from typing import Any, List, Optional, Tuple, Union, Dict, Set, Callable
-from transformers.models.siglip.modeling_siglip import (np)
+
 
 def forward(self, input: Tensor) -> Tensor:
-    return self._conv_forward(input, self.weight, self.bias).to(input.dtype)
+    return self._conv_forward(input, self.weight, self.bias).to(input.dtype).to(input.dtype)
